pyproject.toml
[tool.poetry]
name = "vertex-ai-experiments-easy-sample"
version = "0.1.0"
description = ""
authors = ["f6wbl6"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
google-cloud-aiplatform = "^1.37.0"
kfp = "^2.4.0"
fire = "^0.5.0"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
